pyproject.toml
[tool.poetry]
name = "market-researcher"
version = "0.1.0"
description = ""
authors = ["Richard Blythman <richardblythman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.3"
pydantic = "^2.5.2"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
